requirements.txt
google-generativeai
Pillow
Flask
platformdirs
python-dotenv
